Restore highlighting of distribution page menu items
Request paths are canonicalized to use ; as the query param separator
when comparing the request path to possible active menu paths.  This
canonicalization happens regardless of what the actual request URL uses
for a separator.  Thus, using & as the param separator guarantees a menu
item will never be highlighted as active.