Add distro name regex config, replace explicit CPAN regex
Add the DistributionToQueueRegex config variable, not set by
default. The value used for rt.cpan.org, s/::/-/g, is included in
BugTracker_Config.pm as part of the comment for this variable.

When set, match strings submitted through RT::BugTracker pages and
SimpleSearch.

There is a corresponding commit in RT::BugTracker::Public to migrate
to this config variable from the hardcoded CPAN regex.